created EmployeeSection to be reusable component
