+++ client/src/components/Management/Management.jsx
@@ -1,6 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { useNavigate, useLocation } from 'react-router-dom';
 import MenuSection from '../MenuSection/MenuSection';
+import EmployeeSection from '../EmployeeSection/EmployeeSection';
 
 function Management() {
   const [showDropdown, setShowDropdown] = useState(false);
@@ -15,42 +16,21 @@ function Management() {
     navigate('/calendar', { state: { action, isViewMode: action === 'View' } });
   };
 
-  const token = localStorage.getItem('token');
+  const handleEmployeeSelect = (employee) => {
+    sessionStorage.setItem('clearSession', 'true');
+    navigate('/manage-allergies', {
+      state: {
+        employeeData: {
+          employeeId: employee._id,
+          identity: employee.employeeName,
+        },
+      },
+    });
+  };
 
-  // to fetch employee list
-  useEffect(() => {
-    const fetchEmployees = async () => {
-      try {
-        const response = await fetch(
-          'https://eato-meatplanner.netlify.app/.netlify/functions/employees',
-          {
-            method: 'GET',
-            headers: {
-              'Content-Type': 'application/json',
-              Authorization: `Bearer ${token}`,
-            },
-          }
-        );
-        const responseData = await response.json();
-        // console.log('employee data: ', responseData);
-
-        if (responseData.data) {
-          setEmployees(responseData.data);
-        } else {
-          alert('Unexpected API response format');
-          // console.error('Unexpected API response format:', responseData);
-        }
-      } catch (error) {
-        alert('Error fetching employees');
-        // console.error('Error fetching employees: ', error);
-      }
-    };
-    fetchEmployees();
-  }, []);
-
-  const filteredEmployees = employees.filter((employee) =>
-    employee.employeeName.toLowerCase().includes(searchTerm.toLowerCase())
-  );
+  const handleAddEmployee = () => {
+    navigate('/create-employee');
+  };
 
   return (
     <div className='custom-bg md:h-[995px] lg:h-[907px]'>
@@ -92,84 +72,10 @@ function Management() {
                 ALLERGIES
               </p>
               <div className='flex flex-col justify-center items-center'>
-                <div className='hidden group-hover:flex flex-col w-[320px]'>
-                  <div className='relative w-64'>
-                    <button
-                      onClick={() => setShowDropdown((prev) => !prev)}
-                      className='border-[1px] border-primary text-2xl h-[56px] bg-white flex items-center justify-center gap-6 w-[320px] '
-                    >
-                      <span>Select an Employee</span>
-                      <svg
-                        className='w-4 h-4 text-primary'
-                        xmlns='http://www.w3.org/2000/svg'
-                        fill='none'
-                        viewBox='0 0 24 24'
-                        stroke='currentColor'
-                      >
-                        <path
-                          strokeLinecap='round'
-                          strokeLinejoin='round'
-                          strokeWidth={2}
-                          d='M19 9l-7 7-7-7'
-                        />
-                      </svg>
-                    </button>
-                    {/* dropdown list of employees */}
-                    {showDropdown && (
-                      <div className='absolute mt-1 bg-white border rounded-md shadow-lg z-10 py-2 w-[320px]'>
-                        {/* Search Input Field */}
-                        <input
-                          type='text'
-                          placeholder='Type name here...'
-                          value={searchTerm}
-                          onChange={(e) => setSearchTerm(e.target.value)}
-                          className='py-2 w-[320px]'
-                        />
-
-                        {/* Employee List */}
-                        <div className='max-h-60 overflow-y-auto '>
-                          {filteredEmployees.length > 0 ? (
-                            filteredEmployees.map((employee) => (
-                              <button
-                                key={employee._id}
-                                className='block w-full text-left px-4 py-2 hover:bg-gray-100'
-                                onClick={() => {
-                                  sessionStorage.setItem(
-                                    'clearSession',
-                                    'true'
-                                  );
-                                  navigate('/manage-allergies', {
-                                    state: {
-                                      employeeData: {
-                                        employeeId: employee._id,
-                                        identity: employee.employeeName,
-                                      },
-                                    },
-                                  });
-                                  setShowDropdown(false);
-                                }}
-                              >
-                                <span className=''>
-                                  {employee.employeeName}
-                                </span>
-                              </button>
-                            ))
-                          ) : (
-                            <p className='px-4 py-2 text-gray-500'>
-                              No employees found
-                            </p>
-                          )}
-                        </div>
-                      </div>
-                    )}
-                  </div>
-                  <button
-                    onClick={() => navigate('/create-employee')}
-                    className='border-[1px] border-primary text-2xl h-[56px] bg-white'
-                  >
-                    Add an Employee
-                  </button>
-                </div>
+                <EmployeeSection
+                  onSelectEmployee={handleEmployeeSelect}
+                  onAddEmployee={handleAddEmployee}
+                />
               </div>
             </div>
           </div>
